web/app/requirements.txt
weaviate-client>=4.17.0
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-weaviate>=0.0.5,<0.1.0
transformers>=4.57.1
sentence-transformers>=5.1.2
datasets>=4.3.0
accelerate>=1.11.0
faiss-cpu>=1.12.0
gradio>=5.49.1
python-dotenv>=1.2.1
requests>=2.32.5